Split stressTest to its own file
