+++ packages/test/test-service-load/src/test/mini.spec.ts
@@ -6,13 +6,7 @@
 import assert from "assert";
 import child_process from "child_process";
 
-const childArgs: string[] = [
-	"./dist/nodeStressTest.js",
-	"--driver",
-	"tinylicious",
-	"--profile",
-	"mini",
-];
+const childArgs: string[] = ["./dist/main.js", "--driver", "tinylicious", "--profile", "mini"];
 
 describe("stress test", () => {
 	it("Should return 0", async () => {
